Master

See merge request edusoho/edusoho!15491

+++ src/Biz/Activity/Service/Impl/ActivityLearnLogServiceImpl.php
@@ -3,11 +3,13 @@
 namespace Biz\Activity\Service\Impl;
 
 use AppBundle\Common\ArrayToolkit;
-use Biz\BaseService;
-use Biz\Activity\Service\ActivityLearnLogService;
-use Biz\Activity\Dao\Impl\ActivityLearnLogDaoImpl;
-use Biz\Task\Service\TaskResultService;
 use AppBundle\Common\TimeMachine;
+use Biz\Activity\Dao\ActivityLearnLogDao;
+use Biz\Activity\Service\ActivityLearnLogService;
+use Biz\BaseService;
+use Biz\Crontab\SystemCrontabInitializer;
+use Biz\Task\Service\TaskResultService;
+use Codeages\Biz\Framework\Scheduler\Service\SchedulerService;
 
 class ActivityLearnLogServiceImpl extends BaseService implements ActivityLearnLogService
 {
@@ -17,7 +19,7 @@ class ActivityLearnLogServiceImpl extends BaseService implements ActivityLearnLo
             $data['learnedTime'] = TimeMachine::time() - $data['lastTime'];
         }
 
-        $fields = array(
+        $fields = [
             'activityId' => $activity['id'],
             'userId' => $this->getCurrentUser()->getId(),
             'event' => $eventName,
@@ -25,7 +27,7 @@ class ActivityLearnLogServiceImpl extends BaseService implements ActivityLearnLo
             'learnedTime' => !empty($data['learnedTime']) ? $data['learnedTime'] : 0,
             'data' => $data,
             'createdTime' => time(),
-        );
+        ];
 
         //TODO 临时方案, 要考虑数据的准确性和扩展性
         if (!empty($data['watchTime'])) {
@@ -43,13 +45,13 @@ class ActivityLearnLogServiceImpl extends BaseService implements ActivityLearnLo
     }
 
     /**
-     * @deprecated
-     * @see #isActivityFinished($activityId)
-     *
      * @param $activityId
      * @param $event
      *
      * @return array
+     *
+     * @see #isActivityFinished($activityId)
+     * @deprecated
      */
     public function getMyRecentFinishLogByActivityId($activityId)
     {
@@ -59,14 +61,14 @@ class ActivityLearnLogServiceImpl extends BaseService implements ActivityLearnLo
     }
 
     /**
-     * @deprecated
-     * @see 尽量基于TaskResult统计
-     * @see TaskResultService#calcLearnProcessByCourseIdAndUserId($courseId, $userId)
-     *
      * @param $courseId
      * @param $userId
      *
      * @return array
+     *
+     * @deprecated
+     * @see 尽量基于TaskResult统计
+     * @see TaskResultService#calcLearnProcessByCourseIdAndUserId($courseId, $userId)
      */
     public function calcLearnProcessByCourseIdAndUserId($courseId, $userId)
     {
@@ -76,34 +78,35 @@ class ActivityLearnLogServiceImpl extends BaseService implements ActivityLearnLo
         $learnedTime = 0;
         $learnedTimePerDay = $daysCount > 0 ? $learnedTime / $daysCount : 0;
 
-        return array($daysCount, $learnedTime, $learnedTimePerDay);
+        return [$daysCount, $learnedTime, $learnedTimePerDay];
     }
 
     public function deleteLearnLogsByActivityId($activityId)
     {
-        return $this->getActivityLearnLogDao()->deleteByActivityId($activityId);
-    }
-
-    public function getLastestLearnLogByActivityIdAndUserId($activityId, $userId)
-    {
-        return $this->getActivityLearnLogDao()->getLastestByActivityIdAndUserId($activityId, $userId);
-    }
-
-    public function sumLearnTimeGroupByUserId($conditions)
-    {
-        $results = $this->getActivityLearnLogDao()->sumLearnTimeGroupByUserId($conditions);
-        $results = ArrayToolkit::index($results, 'userId');
-
-        return $results;
-    }
-
-    public function search($conditions, $orderBy, $start, $limit)
-    {
-        return $this->getActivityLearnLogDao()->search($conditions, $orderBy, $start, $limit);
+        $count = $this->getActivityLearnLogDao()->count(['activityId' => $activityId]);
+        if ($count <= 1000) {
+            return $this->getActivityLearnLogDao()->deleteByActivityId($activityId);
+        }
+        $this->getSchedulerService()->register([
+            'name' => "activity_learn_log_delete_job_{$activityId}",
+            'source' => SystemCrontabInitializer::SOURCE_SYSTEM,
+            'expression' => time(),
+            'misfire_policy' => 'executing',
+            'class' => 'Biz\Activity\Job\DeleteActivityLearnLogJob',
+            'args' => ['activityId' => $activityId],
+        ]);
     }
 
     /**
-     * @return ActivityLearnLogDaoImpl
+     * @return SchedulerService
+     */
+    protected function getSchedulerService()
+    {
+        return $this->createService('Scheduler:SchedulerService');
+    }
+
+    /**
+     * @return ActivityLearnLogDao
      */
     protected function getActivityLearnLogDao()
     {
